refactoring vocab to its own file
