+++ src/cript/api/__init__.py
@@ -2,4 +2,4 @@
 
 from cript.api.api import API
 from cript.api.valid_search_modes import SearchModes
-from cript.api.vocabulary_categories import VocabCategories
+from cript.api.vocabulary import VocabCategories
