+++ src/cript/api/vocabulary/__init__.py
@@ -1,0 +1,1 @@
+from cript.api.vocabulary.vocabulary_categories import VocabCategories
